:art: Split Accounts forms in respective domains
