+++ source/YugiAcademy/Accounts/user_profiles/forms/__init__.py
@@ -1,0 +1,1 @@
+from .user_and_profile_form import UserAndProfileForm
