+++ source/YugiAcademy/Accounts/users/forms/__init__.py
@@ -1,0 +1,2 @@
+from .sign_up_form import SignUpForm
+from .update_password_form import UpdatePasswordForm
